increment_counter/requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
cloudevents==1.2.0
colorama==0.4.4
deprecation==2.1.0
Flask==2.0.2
functions-framework==3.0.0
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
proto-plus==1.20.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
watchdog==1.0.2
Werkzeug==2.0.3
